+++ PhpSerializerNET.Test/Deserialize/ArrayDeserialization.cs
@@ -116,5 +116,14 @@ namespace PhpSerializerNET.Test {
 			Assert.AreEqual("World", result[1]);
 			Assert.AreEqual("12345", result[2]);
 		}
+
+		[TestMethod]
+		public void ExplicitToListNonIntegerKey() {
+			var ex = Assert.ThrowsException<PhpSerializerNET.DeserializationException>(() =>
+				PhpSerialization.Deserialize<List<string>>("a:3:{i:0;s:5:\"Hello\";s:1:\"a\";s:5:\"World\";i:2;i:12345;}")
+			);
+
+			Assert.AreEqual("Can not deserialize array at position 0 to list: It has a non-integer key 'a' at element 2 (position 21).", ex.Message);
+		}
 	}
 }